file rearenging and rename
moved linkedList to Model package
rename com.tool.gui to com.tool.view
