--- Moodify/src/main/java/com/tool/Main.java
@@ -1,11 +1,0 @@
-package com.tool;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.printf("Hello and welcome!");
-
-        for (int i = 1; i <= 5; i++) {
-            System.out.println("i = " + i);
-        }
-    }
-}
