+++ Moodify/src/main/java/com/tool/control/MoodShuffler.java
@@ -1,5 +1,7 @@
 package com.tool.control;
 
+import com.tool.model.Node;
+import com.tool.model.DoublyLinkedList;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
