+++ Moodify/src/main/java/com/tool/view/MainFrame.java
@@ -1,6 +1,6 @@
-package com.tool.gui;
+package com.tool.view;
 
-import com.tool.control.DoublyLinkedList;
+import com.tool.model.DoublyLinkedList;
 import javax.swing.*;
 import java.awt.*;
